+++ projects/slormancer-api/src/services/content/index.ts
@@ -22,4 +22,5 @@ export { SlormancerSynergyResolverService } from "./slormancer-synergy-resolver.
 export { SlormancerTemplateService } from "./slormancer-template.service";
 export { SlormancerTranslateService } from "./slormancer-translate.service";
 export { SlormancerUltimatumService } from "./slormancer-ultimatum.service";
-export  { SlormancerValueUpdaterService } from "./slormancer-value-updater.service";
+export { SlormancerValueUpdaterService } from "./slormancer-value-updater.service";
+export { SlormancerAncestralLegacyNodesService } from "./slormancer-ancestral-legacy-nodes.service";